Added multer and new route to handle uploads
Added multer module to handle files posted as a form upon upload of song file. There is a new route that will handle an uploaded file and then send back file data as a confirmation. The data is used to create a button on client side to stream the song.