Playlist, data structure, and song end update
Removed redundant code for creating the playlist on client side, now the playlistAdd function is the only that handles the creation of the song buttons.

Changed the data structure for storing the song button elements from an Array to an Object that has a buton and index field. This makes it easier to get the button and index of a song without having to iterate through an array. The function that previously did that has been removed.

Fixed a bug in the songEnd function that prevented the pause button from switching to the play button when the last song has ended.