Fixed playlist index bug and tweaked songEnd
Fixed a bug from the previous commit involving the change in the data structure for storing song button elements. The index that was being stored for each element was using trying to get the length of overall Object storing the buttons, which was undefined. Switching to using the playlist array length instead fixed it.

createPlaylist function has been removed and the getPlaylist function handles the initial filling of the playlist on launch using playlistAdd function now. 

A tweak was made to the songEnd function so that if someone happens switch songs within a small window of time before the frame requests are stopped and the pause button is swapped, it will no longer stop the new song.