Added handling of identical songs
A song with the same name and extension as an already uploaded song will be considered identical. When such a file is uploaded, it will now replace the file that is currently on the server and a message will be displayed to the client telling them the file was replaced. This will be the method for handling duplicate files for now.