Added song time and fixed replay bug
There is now an element that shows the both the current time and duration of a loaded song. Both are displayed in {minute}:{second} format. Currently it does not handle hours. The timeupdate event fired by the audio element is the basis for updating the current time.

A small tweak was made to the callback function for the canplay event for the audio element in order to fix a bug that toggled the play button one extra time. The issue was caused by canplay being fired again as the entire audio loading is repeated when attempting to replay the same song after it has ended. The play button was thus being toggled once when the user presses the play button and again when the canplay event is fired during the song loading event cycle. This bug caused the audio to start then immediately pause when attempting to replay the last song through the play button after the playlist has ended without repeat active.