Responsive layout and interaction update
Updated the layout to be better responsive. Now appears more organized on smaller width screens like mobile devices and tablets.

Improved the contrast of hover and focus transitions for elements. Also changed font for everything but the title for better visibility.

Tweaked the bar length of the circular visualizer for higher frequencies.