Added visualizer selection menu and some tweaks
A new visualizer menu has been added that allows for user to select their desired visualization on the fly. The visualization will instantly change to the one selected. Layout for the menu has been set for larger displays, but is not yet correct for smaller displays.

Extended the maximum width for which the app will be displayed in the layout for medium sized screens. This was to allow more of the visualizer to be displayed before the right side menus for the large layout move to their position and take up part of the overall display width.

Removed the tweak for the circle visualizer that added extra length to the bars of higher frequencies in an attempt to balance them out with the typically higher amplitude lower frequencies.